+++ pcs/usage.py
@@ -939,9 +939,11 @@ Commands:
         If destination node is specified the resource will be moved to that
         node by creating an INFINITY location constraint to prefer the
         destination node. The constraint needed for moving the resource will be
-        automatically removed once the resource is running on it's new
-        location. The command will fail in case it is not possible to verify
-        that the resource will not be moved back after deleting the constraint.
+        automatically removed once the resource is running on its new location.
+        The command will fail in case it is not possible to verify that the
+        resource will not be moved back after deleting the constraint. If this
+        happens after the location constraint has been created, the constraint
+        will be left in the configuration.
 
         If --strict is specified, the command will also fail if other resources
         would be affected.