+++ pcs_test/tier0/lib/commands/resource/test_resource_move_autoclean.py
@@ -1287,6 +1287,8 @@ class MoveAutocleanFailures(MoveAutocleanCommonSetup):
                     reports.codes.RESOURCE_MOVE_AUTOCLEAN_SIMULATION_FAILURE,
                     resource_id=self.resource_id,
                     others_affected=False,
+                    node=None,
+                    move_constraint_left_in_cib=False,
                 )
             ],
             expected_in_processor=False,
@@ -1318,6 +1320,8 @@ class MoveAutocleanFailures(MoveAutocleanCommonSetup):
                     reports.codes.RESOURCE_MOVE_AUTOCLEAN_SIMULATION_FAILURE,
                     resource_id=self.resource_id,
                     others_affected=True,
+                    node=None,
+                    move_constraint_left_in_cib=False,
                 )
             ],
             expected_in_processor=False,
@@ -1350,6 +1354,8 @@ class MoveAutocleanFailures(MoveAutocleanCommonSetup):
                     reports.codes.RESOURCE_MOVE_AUTOCLEAN_SIMULATION_FAILURE,
                     resource_id=self.resource_id,
                     others_affected=False,
+                    node=None,
+                    move_constraint_left_in_cib=True,
                 )
             ],
             expected_in_processor=False,
@@ -1381,6 +1387,8 @@ class MoveAutocleanFailures(MoveAutocleanCommonSetup):
                     reports.codes.RESOURCE_MOVE_AUTOCLEAN_SIMULATION_FAILURE,
                     resource_id=self.resource_id,
                     others_affected=True,
+                    node=None,
+                    move_constraint_left_in_cib=True,
                 )
             ],
             expected_in_processor=False,