+++ CHANGELOG.md
@@ -21,6 +21,8 @@
   change certain corosync options ([rhbz#2175797])
 - Prevent disabling `auto_tie_breaker` when it would make SBD not working
   properly ([rhbz#2227230])
+- Improved error messages and documentation of `pcs resource move` command
+  ([rhbz#2219554])
 
 ### Changed
 - When exporting constraints in form of pcs commands, constraints containing
@@ -40,6 +42,7 @@
 [rhbz#2175797]: https://bugzilla.redhat.com/show_bug.cgi?id=2175797
 [rhbz#2217850]: https://bugzilla.redhat.com/show_bug.cgi?id=2217850
 [rhbz#2219407]: https://bugzilla.redhat.com/show_bug.cgi?id=2219407
+[rhbz#2219554]: https://bugzilla.redhat.com/show_bug.cgi?id=2219554
 [rhbz#2222744]: https://bugzilla.redhat.com/show_bug.cgi?id=2222744
 [rhbz#2227230]: https://bugzilla.redhat.com/show_bug.cgi?id=2227230